Improved editor and inspector sources
Sources now have a required ID and label. The inspector now has a field for modifying the label and shows the ID in the footer.
Checkbox field type defined for boolean values.
Source fields are better typed.
Created an orderable list ui component for re-ordering sources. Still needs a bit of work but pretty solid so far.