OBS Sync process utilities begun
Sobbing and crying. OBS sync will behind the scenes be considered "obs-ws" for OBS websocket as there are some fundamental things that just can't be done with websockets.
Going to lean more on the remote management and moderation access for now until off the ground. Then can look at a plugin or even fork of the OBS.
There are some breaking changes to how the editor and sources handle and lots of mess with how the obs-ws module handles things, but we're aiming for v1.