Finished up audio meters
Can control volume dB with either number input (in dB) or with a slider that is % scaling. It doesn't apply the log calculations seen in OBS so may be weird at first but hopefully in my head means if someone wants to halve the volume then it's very clear on how to do so.